refactor(frontend): extract small routes to src/pages/ (Phase 2)
Phase 2 of the migration plan in frontend/MIGRATION.md. Creates the
src/pages/ directory (previously missing) and moves the four smallest
routes' logic out of src/routes/, leaving each route file as a thin
entry point per §1 of frontend/CLAUDE.md:

  _authenticated.accounts.tsx        110 -> 6 lines
  _authenticated.connect-with.*.tsx  101 -> 6 lines
  auth.callback.tsx                   49 -> 11 lines
  login.tsx                           87 -> 11 lines

Hoists helpers that were declared inside route files to their own
modules (§6): GoogleIcon + FeatureItem -> src/components/login/;
AccountSection -> src/components/accounts/.

Replaces the four inline .filter() calls on useAccounts() in the
accounts route with three useAccounts(select) calls — the canonical
demonstration case from §3 (each select result is memoized per
subscriber; TanStack Query dedupes on the shared queryKey).

Extracts the useEffect from auth.callback.tsx into a named hook
useRedirectOnAuthSuccess (§4: zero useEffect in components/pages).

Moves login.module.css alongside its owning page component.
