+++ frontend/src/components/login/FeatureItem.tsx
@@ -1,0 +1,9 @@
+import { Text } from '@mantine/core'
+
+export function FeatureItem({ emoji, text }: { emoji: string; text: string }) {
+  return (
+    <Text c="dimmed" size="sm">
+      {emoji} {text}
+    </Text>
+  )
+}
