+++ frontend/src/routes/login.tsx
@@ -1,6 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import { Stack, Title, Text, Button, Box } from '@mantine/core'
-import classes from './login.module.css'
+import { LoginPage } from '@/pages/LoginPage'
 
 export const Route = createFileRoute('/login')({
   validateSearch: (search: Record<string, unknown>): { redirect?: string } => {
@@ -10,78 +9,3 @@ export const Route = createFileRoute('/login')({
   },
   component: LoginPage,
 })
-
-function GoogleIcon() {
-  return (
-    <svg width="18" height="18" viewBox="0 0 18 18" xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M17.64 9.205c0-.639-.057-1.252-.164-1.841H9v3.481h4.844a4.14 4.14 0 0 1-1.796 2.716v2.259h2.908c1.702-1.567 2.684-3.875 2.684-6.615Z"
-        fill="#4285F4"
-      />
-      <path
-        d="M9 18c2.43 0 4.467-.806 5.956-2.18l-2.908-2.259c-.806.54-1.837.86-3.048.86-2.344 0-4.328-1.584-5.036-3.711H.957v2.332A8.997 8.997 0 0 0 9 18Z"
-        fill="#34A853"
-      />
-      <path
-        d="M3.964 10.71A5.41 5.41 0 0 1 3.682 9c0-.593.102-1.17.282-1.71V4.958H.957A8.996 8.996 0 0 0 0 9c0 1.452.348 2.827.957 4.042l3.007-2.332Z"
-        fill="#FBBC05"
-      />
-      <path
-        d="M9 3.58c1.321 0 2.508.454 3.44 1.345l2.582-2.58C13.463.891 11.426 0 9 0A8.997 8.997 0 0 0 .957 4.958L3.964 7.29C4.672 5.163 6.656 3.58 9 3.58Z"
-        fill="#EA4335"
-      />
-    </svg>
-  )
-}
-
-function LoginPage() {
-  const { redirect: redirectTo } = Route.useSearch()
-
-  function handleGoogleLogin() {
-    const apiUrl = import.meta.env.VITE_API_URL ?? 'http://localhost:8080'
-    const url = new URL(`${apiUrl}/auth/google`)
-    if (redirectTo) url.searchParams.set('redirect', redirectTo)
-    window.location.href = url.toString()
-  }
-
-  return (
-    <Box className={classes.container}>
-      <Stack align="center" gap="xl" className={classes.content}>
-        <Stack align="center" gap="xs">
-          <Text className={classes.emoji}>💰</Text>
-          <Title order={1} className={classes.title}>
-            Finança a dois
-          </Title>
-          <Text className={classes.subtitle}>
-            Gerencie as finanças do casal de forma simples e transparente
-          </Text>
-        </Stack>
-
-        <Stack align="center" gap="md" className={classes.features}>
-          <FeatureItem emoji="📊" text="Visão compartilhada das despesas" />
-          <FeatureItem emoji="🔄" text="Transações recorrentes automáticas" />
-          <FeatureItem emoji="🎯" text="Controle de metas juntos" />
-        </Stack>
-
-        <Button
-          size="lg"
-          variant="default"
-          leftSection={<GoogleIcon />}
-          onClick={handleGoogleLogin}
-          className={classes.googleButton}
-          fullWidth
-        >
-          Entrar com Google
-        </Button>
-      </Stack>
-    </Box>
-  )
-}
-
-function FeatureItem({ emoji, text }: { emoji: string; text: string }) {
-  return (
-    <Text c="dimmed" size="sm">
-      {emoji} {text}
-    </Text>
-  )
-}
